+++ R/distribution.R
@@ -2,19 +2,21 @@
 
 #' Discrete gamma distribution
 #' 
-#' @param scale scale parameter of a gamma distribution
+#' @param rate rate parameter of a gamma distribution
 #' @param shape shape parameter of a gamma distribution
+#' @param dist_init whether to distribute initial value across subcompartments following this distribution.
 #' @return a Distribution object for simulator
 #' 
 #' @examples
-#' transitions <- list("S -> I" = d_gamma(1, 5))
+#' transitions <- list("S -> I" = d_gamma(rate = 1, shape = 5))
 #' @export
-d_gamma <- function(scale, shape) {
+d_gamma <- function(rate, shape, dist_init = FALSE) {
   distr <- list(
     distribution = "gamma",
-    scale = scale,
-    shape = shape)
-  
+    rate = rate,
+    shape = shape, 
+    dist_init = as.numeric(dist_init)
+    )
   class(distr) <- c("Distribution", class(distr))
   distr
 }
@@ -24,16 +26,19 @@ d_gamma <- function(scale, shape) {
 #' 
 #' @param scale scale parameter of a Weibull distribution
 #' @param shape shape parameter of a Weibull distribution
+#' @param dist_init whether to distribute initial value across subcompartments following this distribution. (default to FALSE, meaning init value is always in the first compartment)
 #' @return a Distribution object for simulator
 #' 
 #' @examples
 #' transitions <- list("I -> D" = d_weibull(0.6, 2))
 #' @export
-d_weibull <- function(scale, shape) {
+d_weibull <- function(scale, shape, dist_init = FALSE) {
   distr <- list(
     distribution = "weibull",
     scale = scale,
-    shape = shape)
+    shape = shape,
+    dist_init = as.numeric(dist_init)
+    )
   
   class(distr) <- c("Distribution", class(distr))
   distr
@@ -42,6 +47,7 @@ d_weibull <- function(scale, shape) {
 #' Discrete exponential distribution
 #' 
 #' @param rate rate parameter of an exponential distribution
+#' @param dist_init whether to distribute initial value across subcompartments following this distribution. (default to FALSE, meaning init value is always in the first compartment)
 #' @return a Distribution object for simulator
 #' 
 #' @examples
@@ -49,10 +55,12 @@ d_weibull <- function(scale, shape) {
 #' 
 #'
 #' @export
-d_exponential <- function(rate) {
+d_exponential <- function(rate, dist_init = FALSE) {
   distr <- list(
     distribution = "exponential",
-    rate = rate)
+    rate = rate,
+    dist_init = as.numeric(dist_init)
+  )
   
   class(distr) <- c("Distribution", class(distr))
   distr
@@ -62,16 +70,19 @@ d_exponential <- function(rate) {
 #' 
 #' @param mu location parameter or the ln mean
 #' @param sigma scale parameter or ln standard deviation
+#' @param dist_init whether to distribute initial value across subcompartments following this distribution. (default to FALSE, meaning init value is always in the first compartment)
 #' @return a Distribution object for simulator
 #' 
 #' @examples
 #' transitions <- list("I -> D" = d_lognormal(3, 0.6))
 #' @export
-d_lognormal <- function(mu, sigma) {
+d_lognormal <- function(mu, sigma, dist_init = FALSE) {
   distr <- list(
     distribution = "lognormal",
     mu = mu,
-    sigma = sigma)
+    sigma = sigma,
+    dist_init = as.numeric(dist_init)
+    )
   
   class(distr) <- c("Distribution", class(distr))
   distr
@@ -100,46 +111,25 @@ mathexpr <- function(expr) {
 }
 
 
-#' Fixed transition
-#' 
-#' Define a fixed number of individuals of the left compartment transit to the 
-#' right compartment at every time step
-#'
-#' @param x number of individuals who move from one compartment to another
-#' @return a Distribution object for simulator
-#' 
-#' @examples
-#' transitions <- list("S->I" = constant(10))
-#' @export
+#Fixed transition
 constant <- function(x) {
   distr <- list(
     distribution = "constant",
     constant = x)
-  
+
   class(distr) <- c("Distribution", class(distr))
   distr
 }
 
 
-#' Transition probability
-#' 
-#' A fixed percentage of the left compartment transit to the right compartment 
-#' at every time step
-#' 
-#' @param x a float number between 0 to 1
-#' @return a Distribution object for simulator
-#' 
-#' @examples
-#' transitions <- list("S->I"=transprob(0.8))
-#' @export
-transprob <- function(x) {
-  distr <- list(
-    distribution = "transitionProb",
-    transitionProb = x)
-  
-  class(distr) <- c("Distribution", class(distr))
-  distr
-}
+# transprob <- function(x) {
+#   distr <- list(
+#     distribution = "transitionProb",
+#     transitionProb = x)
+# 
+#   class(distr) <- c("Distribution", class(distr))
+#   distr
+# }
 
 
 #' Nonparametric distribution
@@ -148,56 +138,43 @@ transprob <- function(x) {
 #' 
 #' @param ... a vector of values
 #' @return a Distribution object for simulator
+#' @param dist_init whether to distribute initial value across subcompartments following this distribution. (default to FALSE, meaning init value is always in the first compartment))
 #' 
 #' @examples
 #' transitions <- list("S->I"=nonparametric(0.1, 0.2, 0.5, 0.2))
 #' @export
-nonparametric <- function(...) {
+nonparametric <- function(..., dist_init = FALSE) {
   distr <- list(
     distribution = "nonparametric",
-    waitingTime = c(...))
+    waitingTime = c(...),
+    dist_init = as.numeric(dist_init))
   
   class(distr) <- c("Distribution", class(distr))
   distr
 }
 
 
-#' Multinomial
-#' 
-#' Define a set of probabilities of transition from one compartment to multiple
-#' compartments
-#' ```
-#' "I -> R, D" = multinomial(0.9, 0.1),
-#' "I -> R" = d_gamma(3, 2),
-#' "I -> D" = d_lognormal(2, 0.5)
-#' ```
-#' is equal to
-#' ```
-#' "0.9 * I -> R" = d_gamma(3, 2),
-#' "0.1 * I -> D" = d_lognormal(2, 0.5)
-#' ```
-#'
-#' @param ... a vector of probabilities, must add up to 1
-#' @return a Distribution object for simulator
-#'
-#' @export
-multinomial <- function(...) {
-  distr <- list(
-    distribution = "multinomial",
-    probabilities = c(...))
-  
-  class(distr) <- c("Distribution", class(distr))
-  distr
-}
+# multinomial <- function(...) {
+#   distr <- list(
+#     distribution = "multinomial",
+#     probabilities = c(...))
+#   
+#   class(distr) <- c("Distribution", class(distr))
+#   distr
+# }
+
 
 #' @export
 print.Distribution <- function(x, ...) {
   # Print the name of this distribution
   if (x$distribution %in% c("gamma", "weibull", "exponential")) {
     cat("Discretized", x$distribution, "distribution\n")
-    if (x$distribution %in% c("gamma", "weibull")) {
+    if (x$distribution == "weibull") {
       cat("Scale = ", x$scale, ", Shape = ", x$shape, sep = "")
     } 
+    else if (x$distribution == "gamma"){
+      cat("Rate = ", x$rate, ", Shape = ", x$shape, sep = "")
+    }
     else if (x$distribution == "exponential") {
       cat("Rate = ", x$rate, sep = "")
     }
@@ -246,7 +223,8 @@ distributionToJson <- function(distribution) {
   if (distribution$distribution == "nonparametric") {
     dn <- newJsonKeyPair(key = "distribution", value = "nonparametric")
     wt <- newJsonKeyPair(key = "waitingTime", value = newJsonArray(distribution$waitingTime))
-    contents <- c(dn, wt)
+    kp <- newJsonKeyPair(key = "dist_init", value = distribution$dist_init)
+    contents <- c(dn, wt, kp)
   } 
   else if (distribution$distribution == "multinomial") {
     dn <- newJsonKeyPair(key = "distribution", value = "multinomial")