+++ src/Compartment.h
@@ -12,10 +12,14 @@
 class Compartment {
 private:
     std::string compName;
+    bool competingRisks; // keep track whether current compartment handles outgoing transitions as competing risk
 
     // The length of subCompartments is the maximum length of vector transitionProb in outDistributions
     // store population in each sub compartment
-    std::vector<double> subCompartments;
+    // TODO:
+    // - make subcomartments a nested double array, with dim [nOutCompartments, nMaxDwelltime]
+    // - instantiate population for subCompartments based on outDistribution
+    std::vector<std::vector<double>> subCompartments;
 
     // total: the sum of all subCompartments (i.e. population of this compartment) per iteration/timestep
     std::vector<double> compTotal;
@@ -28,11 +32,14 @@ private:
     std::vector<std::weak_ptr<Compartment>> outCompartments;
     // out distribution corresponding to each outCompartment
     std::vector<std::shared_ptr<Distribution>> outDistributions; 
+    // whether to distribute initial value across subcompartment, same length as outDistributions
+    std::vector<bool> distSubCompInit;
     // out weight corresponding to each outCompartment
     std::vector<double> outWeights;
     
     // store population that move out of each sub compartment in the current iteration
     // values will be reset to 0 then updated while iterating through each out compartment
+    // TODO: update length to be the max length of subCompartments (i.e, max of nMaxDwelltime)
     std::vector<double> outSubCompartments;
     // out population corresponding to each outCompartment in current iteration (sum over all outSubCompartment of that out compartment)
     // values will be reset to 0 then updated while iterating through each timestep
@@ -58,20 +65,28 @@ public:
     std::vector<std::shared_ptr<Distribution>> getOutDistributions();
     std::vector<double> getOutWeights();
 
-    std::vector<double> getSubCompartmentValues() {return subCompartments;};
+    std::vector<std::vector<double>> getSubCompartmentValues() {return subCompartments;};
     std::vector<double> getOutValues() {return outTotals;};
     std::vector<double> getOutSubCompartments() {return outSubCompartments;};
 
     // Setters
-    void addOutDistribution(std::shared_ptr<Distribution>& dist);
+    void addOutDistribution(std::shared_ptr<Distribution>& dist, bool distInit = false);
     void addOutWeight(double weight);
     void addInCompartment(std::weak_ptr<Compartment>& linkedCompIn);
     void addOutCompartment(std::weak_ptr<Compartment>& linkedCompOut);
-    void editOutDistribution(std::string outName, std::shared_ptr<Distribution>& dist);
+    void editOutDistribution(std::string outName, std::shared_ptr<Distribution>& dist, bool distInit = false);
     /**
      * Update compTotal value for current iteration
     */
     void initCompTotal(size_t iter);
+    // Function to normalize outWeight 
+    // used in modelJSON (so for model initialization step only) 
+    void normalizeOutWeights();
+    // TODO: 
+    // Function to distribute initValue of a compartment based on the computed probability distribution 
+    // (instead of initValue all in the fist sub-compartment)
+    // This function will also be used modelJSON only
+    // void distributeInitVal();
 
     // subCompartments and outTotals are set after adding all distributions
     void setLengthSubCompartment();