+++ src/DistributionParametric.h
@@ -8,9 +8,11 @@
 class DistributionParametric: public Distribution {    
 protected:
     std::vector<double> transitionProb;
+    std::vector<double> probDist;
     void calcTransitionProb(std::function<double(double)> cdf_func);  
 public:
     double getTransitionProb(size_t index) override;
+    double getProbDist(size_t index) override;
 };