+++ src/ModelJSON.cpp
@@ -132,6 +132,8 @@ ModelJSON::ModelJSON(nlohmann::ordered_json &initialValues, nlohmann::ordered_js
 
             std::shared_ptr<Distribution> distribution;
 
+            bool dist_init = false;
+
             // Set distribution for the inCompartment
             // If outCompartment is not added, then add the distribution, if it has been added then edit the distribution
             if (distributionConfig["distribution"] == "transitionProb") {
@@ -139,33 +141,48 @@ ModelJSON::ModelJSON(nlohmann::ordered_json &initialValues, nlohmann::ordered_js
                 prob *= Distribution::timeStep;
                 distribution = std::make_shared<DistributionTransitionProb>(prob);
             }
-                // Gamma distribution: parameters are "scale" and "shape"
+                // Gamma distribution: parameters are "rate" and "shape"
             else if (distributionConfig["distribution"] == "gamma") {
-                double scale = distributionConfig["scale"];
+                double rate = distributionConfig["rate"];
                 double shape = distributionConfig["shape"];
-                distribution = std::make_shared<DistributionDiscreteGamma>(scale, shape);
+                distribution = std::make_shared<DistributionDiscreteGamma>(rate, shape);
+                // work around json type cast exception, don't remove
+                double tmp_init_config = distributionConfig["dist_init"];
+                dist_init = (bool) tmp_init_config;
             }
                 // Weibull distribution: parameters are "scale" and "shape"
             else if (distributionConfig["distribution"] == "weibull") {
                 double scale = distributionConfig["scale"];
                 double shape = distributionConfig["shape"];
                 distribution = std::make_shared<DistributionDiscreteWeibull>(scale, shape);
+                // work around json type cast exception, don't remove
+                double tmp_init_config = distributionConfig["dist_init"];
+                dist_init = (bool) tmp_init_config;
             }
                 // Exponential distribution: parameter is "rate"
             else if (distributionConfig["distribution"] == "exponential") {
                 double rate = distributionConfig["rate"];
                 distribution = std::make_shared<DistributionDiscreteExponential>(rate);
+                // work around json type cast exception, don't remove
+                double tmp_init_config = distributionConfig["dist_init"];
+                dist_init = (bool) tmp_init_config;
             }
             else if (distributionConfig["distribution"] == "lognormal") {
                 double mu = distributionConfig["mu"];
                 double sigma = distributionConfig["sigma"];
-                
+            
                 distribution = std::make_shared<DistributionLogNormal>(mu, sigma);
+                // work around json type cast exception, don't remove
+                double tmp_init_config = distributionConfig["dist_init"];
+                dist_init = (bool) tmp_init_config;
             }
                 // Values distribution: parameter is a vector "waitingTime"
             else if (distributionConfig["distribution"] == "nonparametric") {
                 std::vector<double> waitingTime = distributionConfig["waitingTime"];
                 distribution = std::make_shared<DistributionNonparametric>(waitingTime);
+                // work around json type cast exception, don't remove
+                double tmp_init_config = distributionConfig["dist_init"];
+                dist_init = (bool) tmp_init_config;
             }
             else if (distributionConfig["distribution"] == "mathExpression") {
                 std::string expression = distributionConfig["expression"];
@@ -179,9 +196,11 @@ ModelJSON::ModelJSON(nlohmann::ordered_json &initialValues, nlohmann::ordered_js
             // Set distribution for the inCompartment
             // If outCompartment is not added, then add the distribution, if it has been added then edit the distribution
             if (inComp.lock()->isOutCompAdded(outCompName) == false) {
-                inComp.lock()->addOutDistribution(distribution);
+                // TODO: adjust setting whether to distribute initial value here
+                inComp.lock()->addOutDistribution(distribution, dist_init);
             } else {
-                inComp.lock()->editOutDistribution(outCompName, distribution);
+                // TODO: also adjust setting whether to distribute initial value here
+                inComp.lock()->editOutDistribution(outCompName, distribution, dist_init);
             }
 
             // If outCompartment has not been added: set linked compartment in, out, outName, weight
@@ -201,6 +220,13 @@ ModelJSON::ModelJSON(nlohmann::ordered_json &initialValues, nlohmann::ordered_js
             comp->addOutDistribution(transitionProb);
             comp->addOutWeight(1);
         }
+
+        // normalize outWeight in case weights does not sum up to 1
+        // also add outWeight in scenarios where outweight is not explicitly defined 
+        comp->normalizeOutWeights();
+        // TODO: also distribute starting value based on specified distribution
+        // TODO: in case of multinomial, create separate chains of subcompartments
+
         comp->setLengthSubCompartment();
         comp->setOutValues();
     }