+++ NEWS.md
@@ -1,5 +1,9 @@
+# denim 1.1.0
+* Fix multinomial vs competing risks transition
+* Add distribute initial value feature
+
 # denim 1.0.1
+* Remove deprecated syntax in C++
 
 # denim 1.0.0
-
 * Initial CRAN submission.